+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ContractStateChange.java
@@ -25,9 +25,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
+ * @deprecated - User mirror nodes for contract traceability instead
+ *
  * The storage changes to a smart contract's storage as a side effect of the function call.
  * See <a href="https://docs.hedera.com/guides/docs/hedera-api/smart-contracts/contractcalllocal#contractstatechange">Hedera Documentation</a>
  */
+@Deprecated
 public class ContractStateChange {
     public final ContractId contractId;
     public final List<StorageChange> storageChanges;
@@ -43,51 +46,51 @@ public class ContractStateChange {
         this.storageChanges = storageChanges;
     }
 
-    /**
-     * Create contract stage change object from protobuf.
-     *
-     * @param stateChangeProto          the protobuf
-     * @return                          the contract stage change object
-     */
-    static ContractStateChange fromProtobuf(com.hedera.hashgraph.sdk.proto.ContractStateChange stateChangeProto) {
-        List<StorageChange> storageChanges = new ArrayList<>(stateChangeProto.getStorageChangesCount());
-        for (var storageChangeProto : stateChangeProto.getStorageChangesList()) {
-            storageChanges.add(StorageChange.fromProtobuf(storageChangeProto));
-        }
-        return new ContractStateChange(ContractId.fromProtobuf(stateChangeProto.getContractID()), storageChanges);
-    }
-
-    /**
-     * Create contract stage change object from byte array.
-     *
-     * @param bytes                     the byte array
-     * @return                          the contract stage change object
-     * @throws InvalidProtocolBufferException       when there is an issue with the protobuf
-     */
-    public static ContractStateChange fromBytes(byte[] bytes) throws InvalidProtocolBufferException {
-        return fromProtobuf(com.hedera.hashgraph.sdk.proto.ContractStateChange.parseFrom(bytes));
-    }
-
-    /**
-     * Create the protobuf.
-     *
-     * @return                          the protobuf representation
-     */
-    com.hedera.hashgraph.sdk.proto.ContractStateChange toProtobuf() {
-        var builder = com.hedera.hashgraph.sdk.proto.ContractStateChange.newBuilder()
-            .setContractID(contractId.toProtobuf());
-        for (var storageChange : storageChanges) {
-            builder.addStorageChanges(storageChange.toProtobuf());
-        }
-        return builder.build();
-    }
-
-    /**
-     * Create the byte array.
-     *
-     * @return                          the byte array representation
-     */
-    public byte[] toBytes() {
-        return toProtobuf().toByteArray();
-    }
+    // /**
+    //  * Create contract stage change object from protobuf.
+    //  *
+    //  * @param stateChangeProto          the protobuf
+    //  * @return                          the contract stage change object
+    //  */
+    // static ContractStateChange fromProtobuf(com.hedera.hashgraph.sdk.proto.ContractStateChange stateChangeProto) {
+    //     List<StorageChange> storageChanges = new ArrayList<>(stateChangeProto.getStorageChangesCount());
+    //     for (var storageChangeProto : stateChangeProto.getStorageChangesList()) {
+    //         storageChanges.add(StorageChange.fromProtobuf(storageChangeProto));
+    //     }
+    //     return new ContractStateChange(ContractId.fromProtobuf(stateChangeProto.getContractID()), storageChanges);
+    // }
+    //
+    // /**
+    //  * Create contract stage change object from byte array.
+    //  *
+    //  * @param bytes                     the byte array
+    //  * @return                          the contract stage change object
+    //  * @throws InvalidProtocolBufferException       when there is an issue with the protobuf
+    //  */
+    // public static ContractStateChange fromBytes(byte[] bytes) throws InvalidProtocolBufferException {
+    //     return fromProtobuf(com.hedera.hashgraph.sdk.proto.ContractStateChange.parseFrom(bytes));
+    // }
+    //
+    // /**
+    //  * Create the protobuf.
+    //  *
+    //  * @return                          the protobuf representation
+    //  */
+    // com.hedera.hashgraph.sdk.proto.ContractStateChange toProtobuf() {
+    //     var builder = com.hedera.hashgraph.sdk.proto.ContractStateChange.newBuilder()
+    //         .setContractID(contractId.toProtobuf());
+    //     for (var storageChange : storageChanges) {
+    //         builder.addStorageChanges(storageChange.toProtobuf());
+    //     }
+    //     return builder.build();
+    // }
+    //
+    // /**
+    //  * Create the byte array.
+    //  *
+    //  * @return                          the byte array representation
+    //  */
+    // public byte[] toBytes() {
+    //     return toProtobuf().toByteArray();
+    // }
 }