+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.17.0
+
+### Added
+
+ * `PrngThansaction`
+ * `TransactionRecord.prngBytes`
+ * `TransactionRecord.prngNumber`
+
+### Deprecated
+
+ - `ContractFunctionResult.stateChanges` - Use mirror node for contract traceability instead
+ - `ContractStateChanges`
+ - `StorageChange`
+
 ## v2.16.3
 
 ### Added